Howdy! Voxel Section Editor III 1.37 was released few hours ago, although it is still pending announcement. Anyway, this little update on the SVN only affects the documentation. The Welcome.txt and Readme.rtf that comes with the program.
Remember that you can checkout this SVN to have the first look on the same internal alpha version of the program that I have in my computer and post your bug reports at VXLSE III Bug Report forum at PPM. In this case, make sure you inform me the number revision where the bug happened. If it's not a recent revision, I'll certainly ignore your bug report and move it to crap forum.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@9 51f8c993-4936-0410-a6d7-a9c99ad3dd25