This revision is almost a release candidate. It brings some new files that are required to make it run properly:
- images/pause.bmp
- images/play.bmp

These images are part of its latest feature:

-> You can now preview HVA animations!

This version also reduces crashes that happens when you switch from Whole Voxel mode to voxels with only one section, although you may still get a crash when you close the application (which is why this isn't the final release).

I'm still looking for the cause of this crash and once it is sorted, 1.38 should be released.

In another note, the Tangent Plane autonormals had its version updated to 8.2, but it is the same thing from the last revision (which I forgot to update the version).

git-svn-id: svn://svn.ppmsite.com/voxel_tools@22 51f8c993-4936-0410-a6d7-a9c99ad3dd25