This is the first alpha from 1.4x series. For this reasons, a lot of changes were done in the file structure of the SVN and since this will just be an update from the old 1.3x series, There is a new directory for 1.4x series, although it's not clear at the moment I'm posting this message on how the program will deal with it.
Anyway, coding wise, the main change was the palette.pas. It's the second generation of palette.pas, now with a proper class for the palette. For this reason, many other files of the program were edited.

For the user, the only possible improvement is the voxel texture tool using the SHP Builder's RGB Advanced Difference Structuralis algorithm to get the voxel colour. Other than this, it is easier to find new bugs rather than improvements in the program.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@25 51f8c993-4936-0410-a6d7-a9c99ad3dd25