The last version was somewhat broken. This one compiles. Also, added RemoveModel on Actor.pas.
Next in the list is to create a ModelBank, VoxelBank, HVABank and DocumentBank. Then, I'll be able to write proper methods to create Actors and Models.

For the end user, this revision is equivalent to revision 61 for every program from this repository. New features will still take few revisions to go online.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@68 51f8c993-4936-0410-a6d7-a9c99ad3dd25