Added a Voxel Bank, which will manage all opened voxels, including cases where the same voxel is used in more than one variable and it should delete them when they are supposed to be deleted. The class exists, but it isn't used yet.
Another thing added is LOD (Level of Detail), which is a way to futurely suppport Renegade models, allowing more or less detailed meshes to be stored in the same model and the render to quickly swicth them (although the swicth itself isn't coded properly).

There are other several random functionalities added.

For the end user, this version still has no differences. Maybe it uses more RAM than the older ones.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@69 51f8c993-4936-0410-a6d7-a9c99ad3dd25