The second red button wasn't pressed yet, so the new rendering code wasn't implemented yet.
The changes from this revision are simple. One is the ability to change the remappable colour from a model and an actor. The second thing is a fix to spelling of the word 'default' in a variable and one function in Voxel_Engine.pas.

For the end user, this revision is equal to the last one.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@89 51f8c993-4936-0410-a6d7-a9c99ad3dd25