Warning: do not run this specific revision. I'm just committing what I have here because I need to transfer it to my laptop and develop it at the university.
In this revision, I'll be very surprised if the render works, but I'm sure it won't work, specially the 3D Preview window.

Anyway, to the changes at the rendering engine:
- Support for screenshots and animations in Bitmap, JPG, GIF, TGA and PNG. That includes screenshots, 360 rotation videos and random animation videos.
- The engine now uses the same trick Stucuk implemented on OS:VV, where it caches the generate picture in a texture, speeding up FPS like freak.
- The engine can render scene on colours and normals mode, just like in normal VXLSE III, where you set spectrum.


At VXLSE III, I've started to implement the new rendering system on both FormMain and Form3DPreview, but I did not finish it yet, which is why this revision is screwed up.

Please, use the previous revision (89) if you are an end user (go to View Log and choose 89 there at svn.ppmsite.com) or be a bit more patient. Thanks in advance.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@90 51f8c993-4936-0410-a6d7-a9c99ad3dd25