End user: DO NOT USE THIS REVISION. It is very unstable.
In this revision, the 3D Preview Window is working again. When I originally coded the new system on it, I forgot to declare the camera. But this is no longer a problem.

I recommend all end users to use revision 57. If you are at svn.ppmsite.com, go to View Log, and click in the 57. From there, you go to vxlseiii4x/vxlse_III.exe and then replace your vxlse_III.exe from a previous VXLSE III instalation with this one.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@94 51f8c993-4936-0410-a6d7-a9c99ad3dd25