This revision fixes several HVA related problems with the addition and removal of new sections. The bug existed on previous versions of the program that had the 3D view with whole unit look, but it became more critical with the new document engine that saves HVA files when you save the voxel.
The 3D view still has a weird bug that mess up the colours of the new section when it is being edited. I don't know yet what is causing it and I'll work on it tomorrow.

For the end user, this revision is stable and recommended, as long as you do not create or remove sections. But even for this purpose, this is still recommended over older versions.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@121 51f8c993-4936-0410-a6d7-a9c99ad3dd25