This revision has some improvements in terms of Triangle Based 3D Model feature. I've remade the concept of semi-surfaces and also fixed a couple of things. However I'm committing this revision for backup purposes only. I'll rewrite the way it finds vertexes and also a system to prevent it from getting some edges.
The results aquired with this feature have several conceptual problems that is seriously limitting it with a couple of holes when we have edges with only one of the vertexes. Also when we have 3 or more faces with neighbors, it causes some weird shapes. And there are a couple of other problems as well with the current concept. Fixing it demands a recode.

For the end user, this revision is stable and recommended, although Rendering model as Triangle Based 3D model is not working as it should yet, although it is getting somewhat closer to the final expected result.


git-svn-id: svn://svn.ppmsite.com/voxel_tools@172 51f8c993-4936-0410-a6d7-a9c99ad3dd25