I've added two new features in the Mesh Effects section of the 3D Preview window in this revision. The first one, is the ability to renormalize the mesh completely ignoring the normals that you've used in the voxel. It sets the perfect normals for the mesh that is being displayed.
The second feature is the Lanczos Smooth, which is based in the Lanczos interpolation algorithm used in many 2D image editing programs such as the GIMP. My Lanczos is adapted, since I'm not dealing with pixels in this feature. I'm working with vertexes. Unlike the other smooth effects, this one only works correctly at the first application. After that, it starts to destroy the model. The result is quite funny.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@182 51f8c993-4936-0410-a6d7-a9c99ad3dd25