This revision has a new option at the 3d Preview window called 'Save Model As'. Surprisingly, this is the option to export the model as obj. Unfortunately, this is really very early efforts. As far as I could test on most obj viewers, they all failed to load these models, except for one, which messed up a couple of vertexes. In short, for the final user, this option is still useless.
The models generated by this feature will be huge in terms of file size. And I really mean it. At the moment, the generated model receives no optimization whatsoever. It generated a bunch of useless vertexes and normals that could be removed by simply detecting that their normal would be (0,0,0). The polygons with messed up normals usually have normals like composed with 0, 1 and -1 (even after you use ReNormalize Model feature.

Also, it doesn't save materials, nor try to use inexistent ones, so I believe that it also helps to screw up a couple of obj viewers. Another limitation is that it exports quads instead of triangles, which also screws up a couple of external programs.

In short, don't use it yet, but expect it to become useful in the next revisions.


In a short note: The model I've tested was the hmec.vxl (Mammoth Mark II), which has about 16 sections. The obj file size was 1.63mb, which is really a lot for an obj file. I've tested with Blender, Brownz 3D and Obj Viewer from http://www.aweapp.com/obj_viewer which was the only application that opened the .obj file (with messed up vertexes and faces).


git-svn-id: svn://svn.ppmsite.com/voxel_tools@185 51f8c993-4936-0410-a6d7-a9c99ad3dd25