This revision has a lot of changes. It starts with the 3D Modelizer tool, inside Tools. This is another 3D viewer, except that it is totally focused at exporting the model that you already have done. Whatever you change in the model once this tool is opened, it won't affect it. So, all the effects, such as smooths, unsharping mask, deflate and inflate are here.
The 3D Preview window has lost some features, compared to the earlier revisions. The model effects were removed. However, you have a new model quality based in quads available, that is actually cubes with lanczos smoothing. And it still works quite quickly when you are editing the model.

There are other changes and fixes. The Save Model As feature (available on both 3D preview window and 3D Modelizer) is working better and it will no longer generate invalid faces, neither faces with invalid vertexes. Unfortunately, I still haven't added the colours.

There are many other changes in the core of the rendering engine. FPS limits will now affect new rendering window, among other things. Unfortunately, there is a bug that I couldn't figure out how what causes it. If you open a new window (3d preview or modelizer) and modify the model from the previous window (i.e.: change remappable), there is a possibility that the unit in this new window will disappear. It seems to be related to HVA, since I've seen the position of the model sections to be twisted in few occasions. But then, I've separated the HVA from the modelizer and it still had this problem. It looks like some messed up object assign and I'll have to look carefully into the code.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@186 51f8c993-4936-0410-a6d7-a9c99ad3dd25