This revision comes with some interesting changes and one new feature.
The main change is related to the custom schemes code, which was almost entirely rewritten. We have a new CustomScheme.pas with a class that reads these kind of files and save them as well, although the save function is not used yet (and not tested either). With this change, the program starts much faster, executes colour scheme replacements quicker and I've added a neat feature to update the colour schemes list. So, you don't need to restart the program to get it updated.

I've also fixed a couple of memory leaks related to custom schemes and I've also fixed a problem where the program could crash when you exit it.

The menu has an interesting, yet disabled upcoming feature which is my next target. One of the main features from the VXLSE III 1.40. But it will use an external program to perform this feature, which you need to download it outside. Tomorrow, the following url will be available: http://get3ds2vxl.ppmsite.com

Can you feel what's coming next?

git-svn-id: svn://svn.ppmsite.com/voxel_tools@225 51f8c993-4936-0410-a6d7-a9c99ad3dd25