This revision has the neat feature implement: It fully interacts with ViPr's tool, using it to import .3ds models as voxels. I think it also supports .md3 or something from Doom3... but it's not as reliable as .3ds stuff.
Make sure you've enabled the voxelizer and you must not use the Batch Converter, otherwise the program won't be sure that you've actually generated a voxel with it.

If you don't have 3ds2vxl, grab it from http://get3ds2vxl.ppmsite.com

Once ViPr makes a new home for it, I'll redirect it to it.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@226 51f8c993-4936-0410-a6d7-a9c99ad3dd25