WARNING: THIS REVISION IS CONSIDERED UNSTABLE. WE RECOMMEND USING REVISION 227 IF YOU DON'T WANT TO HAVE ANY RISKS.
This revision fixes the .OBJ export coordinates and changes the entire coordinates system of the program (again).

X -> Horizontal/Lateral
Y -> Height
Z -> Depth

Exactly like OpenGL.

The original voxel loading function loaded X as depth, Y as lateral and Z as height. The program code was completely messed up because of that. Now it converts the Westwood coordinates into OpenGL when it loads a voxel. The user interface is all based on OpenGL now. The conversion back to Westwood coordinates happens only when saving voxel and HVA files and it converts back to OpenGL before the user gets control of it again.

The program is still saving and loading voxels and HVAs correctly, as far as | could test. The unstability comes from the Nugde feature which is weird at the 3D view with some views (back and top mainly). The mirror front to back and back to front is not working as it should.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@228 51f8c993-4936-0410-a6d7-a9c99ad3dd25