This revision adds a functional (yet limited and not much reliable) DDS saving function at DDS.pas. The user will be able to save the screenshot of the Modelizer view in the DDS format, but I definitely do not recommend it at all. It simply saves the texture generated to cache the screen in OpenGL, which distorted dimensions (all of them are power of 2).
The DDS files it generates are compressed with DXT5 format, from DirectX. It directly uses the video card to compress it and, for this reason, it flips the texture upside down. I have no clue if it will be accepted by video cards that lack support to Open GL 2.0 or higher. The Windows Texture Viewer (http://developer.nvidia.com/object/windows_texture_viewer.html) is able to open them without problems. GIMP plugin could not open it. Paint Shop Pro with the nVidia plugin opens it correctly. I did not test it with Photoshop, because I do not have it here.

So, what's the matter of having a useless 'please do not use me' save DDS function? We can save compressed textures with mipmaps and the program is almost ready to generate these textures.

- This revision fixes and improves the coding of GetPow2Size function at BasicFunctions.pas.
- This revision fixes a couple of conceptual problems from the GetTextureCoordinates function at the ClassTextureGenerator.pas.

For some unknown reason, if I try to free a TDDSImage object, the program crashes. I have no clue why does it happen, since this class doesn't really hold any variables at the moment.

I'll try to improve its code later.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@239 51f8c993-4936-0410-a6d7-a9c99ad3dd25