We finally have an update. And this one is big and may affect a lot of users.
First of all: VXLSE III now officially requires OpenGL 2.0

The reason is the shader support. Making the phong shader work correctly was a pain in the ass, but hopefuly, new shaders should be much easier to create, with a fully working sample.

This version also fixes a bug where, if you had multiples 3d views opened, resizing one of them would affect the other window.


Finally, it has a draft of the next features: the auto-repair and auto-update. So, if you can't be bothered to download the shaders on your own (which you must have to run the program correctly), wait the next revision and once you run it, it will automatically download it for you.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@244 51f8c993-4936-0410-a6d7-a9c99ad3dd25