This revision gets even better results with the Generate Diffuse Texture option at the Modelizer Windows -> Texture Effects.
I'm very close to my final expectations. I just need to fix the border of each partition, which is distorting the colours. Then, I'll have to merge all sections into a single texture image, but that's it, since the existing code that merges partitions could be adapted to do that as well.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@253 51f8c993-4936-0410-a6d7-a9c99ad3dd25