This revision generates good and satisfactory textures for single section models. I've fixed the remaining bilinear filtering issues that generated black or white borders (depending on the revision you used to have) in the texture and I've optimized several parts of the algorithm, which makes the texture generation considerably faster now.
I've started to code a way to export material files with OBJ, but I'm not done with that yet.

Before I finish it, I'll have to change the way textures are generated, in order to generate a single diffuse texture for all sections. So, that's my number 1 priority right now.

Then, we'll have a proper OBJ export with textures.

VXLSE III is slowly being transformed into a 3D modelling tool...

git-svn-id: svn://svn.ppmsite.com/voxel_tools@261 51f8c993-4936-0410-a6d7-a9c99ad3dd25