In the last revision, I forgot to commit the ClassIntegerSet.pas file, so this time I'm including it, as well as few other files.
In this revision, the Mesh Optimization feature from the 3D Modelizer window has been rewritten and received several fixes, working correctly (or better) when the model has textures. I've removed its code from Mesh.pas and I've created an entire new class to deal with it: the TMeshOptimizationTool.

The end user will notice that when you optimize the mesh of models with textures, the texture coordinates no longer get as messed up as it used to happen. The result is visually good, except when the mesh gets too much deformed. The next step to improve the mesh optimization tool will happen once we start to deal with normal maps. For now, we can remove the colour restriction, add some stronger angles like 30 or 40 degrees and get some interesting results in some models, reducing 50% or more faces of the mesh.


The last included file is a ColladaFile.pas, which should allow models to be exported in the collada (.dae) format, which holds much more details than .obj and it is accepted in most 3D modelling programs. But the code in this file is in very early stages.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@268 51f8c993-4936-0410-a6d7-a9c99ad3dd25