This revision speeds up a bit most operations in the 3D Modelizer window, such as mesh smoothing, texture generation, mesh optimization, among others. I've done a couple of tweaks to the ClassNeighborDetector.pas, which is used by these procedures.
Also, the shaders are re-enabled, because I need to figure out which gaps are generated by the texture alpha channel and which ones are real gaps.

Finally, the ClassMeshOptimizationTool.pas has few new attempts to improve the final quality of the optimized mesh, although they are disabled in this build, the code is there for testing.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@277 51f8c993-4936-0410-a6d7-a9c99ad3dd25