This revision will be very useful for those who have been experiencing problems with the shaders used in the beta revisions of this program. The main preview and the 3D preview window are no longer using shaders. The Modelizer has received the option to disable and enable them whenever you want. At this moment, I must admit that the models look (much) better with disabled shaders. For some weird reason, several textured models with shader are looking transparent. So, I'll have to fix it too.
And, I've also changed the way it converts Face Normals into Vertex Normals, however, there is no visible difference in the result. Surprisingly enough, the black vertexes that results from this conversion only appears when I use shader. If I disable it, the quality of the normals is perfect. So, there is a good chance that my shader is problematic rather than the Face to Vertex Normals conversion.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@291 51f8c993-4936-0410-a6d7-a9c99ad3dd25