This revision is a work in progress code that I need to transfer to my laptop. But, once it's fully coded, it should bring a very interesting feature for the end user: The ability to check if the normals are correct, but representing each normal vector as a line outside the model. With these lines, we'll be able to see if our normals makes any sense or are lunatic.
This feature is actually challenging in terms of software engineering, since it actually creates a mesh inside another mesh. However, after some thought, I've decided to treat it as a plugin, rather than using a mesh inside another mesh. Afterall, it is a kind of data that shouldn't be exported by the end user and it can be enable or disabled by user request. So, that's why we have a template for Mesh Plugins called MeshPluginBase.pas and the NormalsMeshPlugin.pas which is the plugin to preview the normals. The only problem is that, if plugins can be rendered, I'll need to split the rendering code from the Mesh, which is why we have the RenderingMachine.pas.

But, as I said earlier, this feature is not implemented yet. So, this revision brings no news at all for the end user. I'm sorry.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@294 51f8c993-4936-0410-a6d7-a9c99ad3dd25