This revision has an early beta version of the Display Normal Vectors feature at the Modelizer window. It's in the Display section of the popup button.
This feature shows a line representing the direction of each normal (from face or vertex, depending on how normals are set in your model). Its direction comes from the dark to the light red.

This feature still have a lot of problems that needs to be fixed. Sometimes, it won't show up the normals from all vertexes from the model. There is another situation where they disappear depending on the angle and zoom from the camera. I don't know exactly what is causing these things to happen, but I wanna fix them soon and bring this feature to the other previews.   

git-svn-id: svn://svn.ppmsite.com/voxel_tools@297 51f8c993-4936-0410-a6d7-a9c99ad3dd25