After some thoughts, I've decided to let the shader ignore the material properties when it doesn't use textures. This shouldn't be really a problem, since material will only be useful on models with textures... and the shader that uses diffuse textures works well with materials.
The end user may need to download the latest shaders to make use of the changes of this revision. The quickest way to do that is to make sure the program is not running. Then, open your windows explorer, go to the VXLSE III directory and then, go to the shaders directory. Then, delete every file from it and run the program.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@303 51f8c993-4936-0410-a6d7-a9c99ad3dd25