Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems once you generate texture on the 3D modelizer window. (If you renormalize mesh, for instance, parts of your model will disappear).


Anyway, this revision starts to use the new neighborhood code to find the normals (which are looking better) and process the mesh. It is not used in colours (the code is too old and not priority). I'll soon implement the code in the texture generation and mesh optimization codes.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@317 51f8c993-4936-0410-a6d7-a9c99ad3dd25