Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems when you do operations after you generate texture on the 3D modelizer window. (If you renormalize mesh, for instance, a small part of your model may disappear or not).


This revision hides some of the bugs, although it may add some memory leak related to QuadFace neighborhood, since the QuadFaceNeighbors from NeighborhoodDataPlugin.pas is not closing correctly. It also improves the way meshes are renormalized after a texture is generated, although a single face still disappeared with Demo Truck. Other units may not even be textured and the reason is still unknown.

Anyway, my focus now is to get the damn bump mapping generation procedure working, even if the program won't use this feature for a while. I have less than 24 hours to finish it and I still have no clue on how I'll paint it. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@319 51f8c993-4936-0410-a6d7-a9c99ad3dd25