Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems when you do operations after you generate texture on the 3D modelizer window. (If you renormalize mesh, for instance, a small part of your model may disappear or not).


Most of the bump mapping texture generation code is here already. But I know that this bump mapping will crash because it will try to view Pixels from the Diffuse Bitmap that are out of bounds. But the program is not using it yet. I also need to create a shader for dot3 bump mapping. I'm just committing to save the code. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@320 51f8c993-4936-0410-a6d7-a9c99ad3dd25