Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems when closing the 3D Modelizer window, if you are using a triangle based mesh. We also don't know if Mesh Optimization crashes or not. Neither if it works correctly.


This revision fixes several problems related to ReNormalize Mesh and other Normal operations done in the 3D Modelizer window, which makes it more stable than previous versions. I've also splitted all the code related to Normals in a single file called ClassMeshNormalsTool.pas.

The bump mapping rendering code is still not finished and this is why I'm not testing it (and the normal map generation code).

git-svn-id: svn://svn.ppmsite.com/voxel_tools@322 51f8c993-4936-0410-a6d7-a9c99ad3dd25