Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems when closing the 3D Modelizer window, if you are using a triangle based mesh. Operations done after you use the Mesh Optimization may provide unexpected results.


This revision improves colour operations on the 3D Modelizer window, transferring most of its code to ClassMeshColoursTool.pas.

The bump mapping rendering code is still not finished and this is why I'm not testing it (and the normal map generation code).

git-svn-id: svn://svn.ppmsite.com/voxel_tools@323 51f8c993-4936-0410-a6d7-a9c99ad3dd25