Warning: This version is Unstable. The latest stable revision is 315.
In this revision, you may face problems when closing the 3D Modelizer window or setting a new Model Quality. As far as we could diagnose, there is data being written inside class code (at the moment the victim is TMeshMaterial).

We have a tweak to actually preview the results generated by the bump mapping and normal mapping operations (which are not rendering anything at the moment). It generates bitmap pictures inside your VXLSE III. Note that this is a temporary solution, so we can debug and see what exactly is being generated by our code so far.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@328 51f8c993-4936-0410-a6d7-a9c99ad3dd25