Unlike the previous revisions, this one is actually stable.
However, Bump mapping and Normal mapping generator are not working yet (the model will disappear).

This revisions slow downs the program a bit, but for a honourable reason. The newest neighborhood data plugin that improved the quality of all operations generated too much fragmented memory (because of ClassNeighborDetector.pas) and, THAT really messed up with Delphi's Garbage Collector. If you are making a program that uses too much memory from the heap and store data on it, at some moment, Delphi Garbage Collector will start to crash when you setLength from dynamic arrays that used to work well in your program.

To solve this problem, you need to move this data into a single and big array that you can kill at once with SetLength later. And that's what I did at the ClassNeighborDetector.pas.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@329 51f8c993-4936-0410-a6d7-a9c99ad3dd25