This revision fixes an annoying problem with the Repair Assistant. It will no longer lock the interface when it downloads the program files.
We've also got rid of Indy for once and for all. All our web code is located in a file called Internet.pas, although that may change later.

I've also reorganized the ClassTextureGenerator.pas (which was really messed up), splitting it into 3 files:

- ClassTextureAtlasExtractor.pas holds the code to extract the texture coordinates from a mesh.
- ClassTextureGenerator.pas is now only responsible to generate the texture maps (diffuse, normal and bump).
- ClassTriangleFiller.pas holds the functions that paints triangles and it is used only by ClassTextureGenerator.pas at the moment.


The bump mapping texture generation code is better than it was, but the textures that are being generated so far have nothing to do with bump maps. It's something else... from another world, which I'll investigate and figure out how to make proper bump maps. But all faces are being painted now.

The normal map is still darker than it was supposed to be and I don't know why it is so dark.

There are other numerous code optimizations elsewhere and few interesting functions created at math3d.pas to make the job easier.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@338 51f8c993-4936-0410-a6d7-a9c99ad3dd25