There are a lot things wrong with the current bump mapping algorithm. What it paints doesn't really make sense and I don't know if the shader makes sense either... neither if the tangent and binormal values are being passed correctly. This revision just makes the process a bit faster using less RAM. But I do not trust the current bump mapping at all.
In a short note, in this revision you can extract a 'height map' texture from the model, once you've textured it. It's not really useful for anything, except if you wanna do your own bump mapping texture from the model that you can export with the Modelizer. For me, it is useful to debug the process.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@340 51f8c993-4936-0410-a6d7-a9c99ad3dd25