This revision is completely bugged and I'm just committing because I'm going to travel and I wanna resume my progress. It compiles, but it doesn't render yet. Once it everything gets fixed, the future versions will run slower than the older revisions.
Note: I did not commit the executable, so you can still download the latest vxlse_III.exe.

The reason of this future slower performance is that this revision splits the Geometry from the mesh, allowing multiples geometry types in the same mesh. At the moment, we could have something like a mesh with triangles and quads at the same time. In the future, we could implement meshes with beta splines.

Why does VXLSE III needs it? The models that it generates currently are full of topological problems. They are not manifolds. That's the reason why you can't simplify the models correctly and also the reason why the texture generator divides the model too much and there are some random problems with normals and the shape itself on areas where a voxel has a neighbor linked by an edge or vertex.

So, I'm recoding the whole way it extracts a polygonal model from a voxel model and there is a lot of work to do. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@365 51f8c993-4936-0410-a6d7-a9c99ad3dd25