This revision is stable, although I'm not committing the executable yet (revision 361 is still recommended).
I've fixed a horde of problems with the new model quality that was conceived to extract a topologically correct polygonal model from a voxel model, but it still fails hard and there are still a lot of bugs to fix with it. I don't even know if it is crashing in this exact moment, so it's highly recommended to not use it yet. 

git-svn-id: svn://svn.ppmsite.com/voxel_tools@372 51f8c993-4936-0410-a6d7-a9c99ad3dd25