This revision has a very interesting new feature: Tools -> Quality Analysis -> Topology.
It verifies the quality of the topology of your voxel model. If your score is 100%, it means that you can have perfect normals (if you set them correctly). If it is not (which is what usually happens), you may have pixels with ambiguous normals. Most Westwood voxels I've tested had scores varying between 75% and 95%. Is your voxel good enough? Give it a try!

git-svn-id: svn://svn.ppmsite.com/voxel_tools@402 51f8c993-4936-0410-a6d7-a9c99ad3dd25