We have better results for any operation at the 3D Modelizer window, due to an important bug fix at the ClassNeighborDetector.pas.
This revision also has a feature called "Convert Quads to "4-8 Subdivision" Triangles. It transform a quad into 4 triangles adding a vertex at its center (that is part of the 4 triangles). The great difference of doing that is that the vertexes of the mesh keeps having a uniform distributivity of its neighbor faces (without the need of the crazy data structure that I was using a while ago, also known as QuadFaces at NeighborhoodDataPlugin.pas).

git-svn-id: svn://svn.ppmsite.com/voxel_tools@414 51f8c993-4936-0410-a6d7-a9c99ad3dd25