This revision brings a new method that is being used for mesh processing, that allows us to use filters in a much more accurate way, specially if the mesh does not have a uniform distribution of polygons. So far, we were using a Laplacian operator with a fixed distance for all neighbors, as described at my SBGAMES article. That method gave the same weight for polygons that takes different area of a neighborhood of a vertex. This new method considers this factor and even more. And the results so far are interesting.
You can find them in the 3D Modelizer window, by pressing the magic button -> Model Effects -> Slow and Flexible bla bla bla... I forgot. Anyway, it's not slow here. But the other one is slightly faster. But it's really slightly, because I haven't noticed the difference. I only know that it is faster because of the code.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@422 51f8c993-4936-0410-a6d7-a9c99ad3dd25