This revision eliminates double faces from the Topologically Correct Triangle Models.
I've also added a code to detect the amount of cores that your processor has. At the moment, it doesn't do anything with this information. In the future, it should help the program to run multi-threaded functions and speed up rendering or 3D mesh generation. I hope it doesn't crash your system. And my desktop's processor (AMD Phenom X3 8750) was detected correctly with its 3 cores.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@426 51f8c993-4936-0410-a6d7-a9c99ad3dd25