First of all, I've received a code from HBD of his version of VXLSE III with a lot of modified stuff from a very old version of VXLSE III (committed right after 1.38's release, I guess). He has made a lot of stuff, from a new autonormals, 3D flood fill tool, a complete new 3D preview window with an entirely different way of interacting and he had a preliminary work in progress code to directly paint in the 3D view. I'll slowly port some of these features to the current version of VXLSE III.
So, the changes for this revision are:
- Updates at the Mesh Smooth code (the newest and slowest one) at 3D Modelizer.
- Fixed a bug where a crash would happen if you changed the palette without loading a voxel file.
- HBD's Autonormals has been ported to the program and it is now available at the Auto Normals window. He claims he modified the 6-faced autonormals. I think it is quite different, but it provides very nice results. It works better with range = 1 and smooth = 1. Although you may select non-integer values, at the moment, the algorithm is just trunc the value provided for smooth and range.

In a small note, the ported code from HBD comes from an older version of VXLSE III and it may contain code compatible with the old (and almost deprecated system). I think that functions like CreateSurfaceMap might be unnecessary and the current VXLSE III has several implementations of a 3D flood and fill. Also, a lot of variables were not documented and I still haven't tried the code by hand to figure out what do they do.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@429 51f8c993-4936-0410-a6d7-a9c99ad3dd25