Reformed command system, OpenCL support, improved texture generators and much more.
I'm committing the current version because I need to transfer it to my laptop, so:

This commit is considered unstable due to the absurd amount of changes where several were untested or not totally safe. Be careful with features such as 3D Preview Window (there is a chance of crashing when closing), Mesh Optimization (not tested at all).

This is a major commit which has the first implementation of the new Controller system and Undo system for models only (voxels will get that later). The controller receives and stores commands, which in the future will be used by other users to replicate a model from a certain machine on other machines (like multi-player). The undo/redo system has been coded, but it wasn't implemented yet, which means that, at the moment, it only consumes RAM.

This revision also has support for OpenCL (nVidia and AMD machines only). The first feature that uses OpenCL is the Texture Atlas Generator (Origami), which was tested only on nVidia machines at this moment (CUDA). It has received several fixes and it generates much better results than ever. In some situations (small models with no ramifications), the texture is shorten to a single card, which is excellent. However, there are still a couple of problems that causes borders of the cards to show up in the model. I don't know what causes it yet. The other texture atlas extraction method generates better results due to an algebraic fix on VertexTransformationUtils.pas. The names of the texture atlas extraction methods in the UI were updated.

There is also a debug texture atlas feature that is allowing me to debug the result of both texture atlas extraction methods. The result of that thing is only shown when you export all textures. This should be fixed in a later time.

In a later time, I'll update the code for the non-GPU version of Origami (it might be messed up at the moment) and add it for non nVidia/AMD machines. I'll also add an option to use GPGPU or not at the Preferences.

There were many other code changes and things being reorganized. I also need to delete a bunch of stuff on Actor, Model, LOD and Mesh. And I will eventually split the voxel part from the Actor, Model, LOD and Mesh, so I can reuse this code on other projects that may not necessarily use voxels.

Finally, some files were deprecated and they are here for backup purposes only.

If you are looking for a stable version, keep using the previous revision (441) for now. Sorry for any inconvenience.

git-svn-id: svn://svn.ppmsite.com/voxel_tools@442 51f8c993-4936-0410-a6d7-a9c99ad3dd25